refactor(puzzle): split gather.ts into path-bounds, group-bounds, and layout modules (closes #267)
`src/game/gather.ts` did three semantically distinct things — SVG path
parsing, group bounds primitives, and gather layout. Splitting them
along those seams keeps each file focused on one layer.

- `src/game/path-bounds.ts` — pure SVG `d`-attribute AABB parsing
  (`getPathBounds`).
- `src/game/group-bounds.ts` — `getGroupOffsetBounds`,
  `getGroupLocalBounds`, `getGroupVisualBounds`. `getGroupLocalBounds`
  now imports `getPathBounds` from the new module.
- `src/game/gather.ts` — only the layout algorithm
  (`computeGatheredPositions`, `applyGatheredPositions`,
  `GATHER_PADDING`, `WorldRect`, `GatherResult`); imports
  `getGroupVisualBounds` from `./group-bounds.js`.

Tests are split alongside (`path-bounds.test.ts`, `group-bounds.test.ts`,
`gather.test.ts`). The barrel (`src/game/index.ts`) re-exports all
helpers from their new homes, so `main.ts` and other call sites are
unchanged. `rotate-group.ts` is updated to import from the new
`group-bounds.js` location.

The `getPathBounds` regex loops are converted from
`while ((m = re.exec(s)) !== null)` to `for (const m of s.matchAll(re))`
— same iteration semantics for global regexes, slightly cleaner.
