+++ src/game/gather.test.ts
@@ -1,5 +1,5 @@
 /**
- * Tests for the gather pieces logic.
+ * Tests for the gather pieces layout algorithm.
  */
 
 import { describe, it, expect } from 'vitest';
@@ -7,10 +7,6 @@ import type { Piece, PieceGroup } from '../model/types.js';
 import {
     computeGatheredPositions,
     applyGatheredPositions,
-    getGroupOffsetBounds,
-    getGroupLocalBounds,
-    getGroupVisualBounds,
-    getPathBounds,
 } from './gather.js';
 import { makeRectPiece } from '../test-helpers/fixtures.js';
 
@@ -27,22 +23,6 @@ function makeMultiGroup(
 }
 
 const landscapeAspect = 800 / 600; // 1.33
-
-describe('getGroupOffsetBounds', () => {
-    it('should return zero bounds for a single piece at origin', () => {
-        const group = makeGroup(1, 0, 0);
-        expect(getGroupOffsetBounds(group)).toEqual({ minX: 0, minY: 0, maxX: 0, maxY: 0 });
-    });
-
-    it('should compute correct bounds for multi-piece group', () => {
-        const group = makeMultiGroup(1, { x: 0, y: 0 }, [
-            [0, { x: -50, y: -50 }],
-            [1, { x: 50, y: 0 }],
-            [2, { x: 50, y: 50 }],
-        ]);
-        expect(getGroupOffsetBounds(group)).toEqual({ minX: -50, minY: -50, maxX: 50, maxY: 50 });
-    });
-});
 
 describe('computeGatheredPositions', () => {
     it('should return empty result for no groups', () => {
@@ -112,74 +92,6 @@ describe('computeGatheredPositions', () => {
     });
 });
 
-describe('getPathBounds', () => {
-    it('should return start point bounds for empty path', () => {
-        const bounds = getPathBounds('', { x: 10, y: 20 });
-        expect(bounds).toEqual({ minX: 10, minY: 20, maxX: 10, maxY: 20 });
-    });
-
-    it('should handle absolute line commands', () => {
-        const bounds = getPathBounds('M 0 0 L 100 50 L 50 100', { x: 0, y: 0 });
-        expect(bounds).toEqual({ minX: 0, minY: 0, maxX: 100, maxY: 100 });
-    });
-
-    it('should handle relative line commands', () => {
-        const bounds = getPathBounds('l 100 0 l 0 80', { x: 10, y: 10 });
-        expect(bounds).toEqual({ minX: 10, minY: 10, maxX: 110, maxY: 90 });
-    });
-
-    it('should include cubic bezier control points (absolute)', () => {
-        // Control points at (50, -30) and (50, 130) extend beyond endpoints
-        const bounds = getPathBounds('C 50 -30, 50 130, 100 50', { x: 0, y: 0 });
-        expect(bounds.minY).toBe(-30);
-        expect(bounds.maxY).toBe(130);
-        expect(bounds.maxX).toBe(100);
-    });
-
-    it('should include cubic bezier control points (relative)', () => {
-        const bounds = getPathBounds('c 20 -40, 80 -40, 100 0', { x: 0, y: 50 });
-        expect(bounds.minY).toBe(10); // 50 + (-40)
-        expect(bounds.maxX).toBe(100);
-    });
-
-    it('should include quadratic bezier control points', () => {
-        const bounds = getPathBounds('Q 50 -20, 100 0', { x: 0, y: 0 });
-        expect(bounds.minY).toBe(-20);
-        expect(bounds.maxX).toBe(100);
-    });
-
-    it('should handle H and V commands', () => {
-        const bounds = getPathBounds('H 200 V 150', { x: 10, y: 10 });
-        expect(bounds).toEqual({ minX: 10, minY: 10, maxX: 200, maxY: 150 });
-    });
-
-    it('should handle relative h and v commands', () => {
-        const bounds = getPathBounds('h 50 v 30', { x: 10, y: 10 });
-        expect(bounds).toEqual({ minX: 10, minY: 10, maxX: 60, maxY: 40 });
-    });
-
-    it('should handle S (smooth cubic) commands', () => {
-        const bounds = getPathBounds('S 50 -20, 100 0', { x: 0, y: 0 });
-        expect(bounds.minY).toBe(-20);
-        expect(bounds.maxX).toBe(100);
-    });
-
-    it('should handle Z command without error', () => {
-        const bounds = getPathBounds('L 100 100 Z', { x: 0, y: 0 });
-        expect(bounds.maxX).toBe(100);
-        expect(bounds.maxY).toBe(100);
-    });
-
-    it('should handle a realistic jigsaw tab path', () => {
-        // Simulates a tab that bulges outward (negative y = upward)
-        const path = 'L 30 0 C 35 -25, 65 -25, 70 0 L 100 0';
-        const bounds = getPathBounds(path, { x: 0, y: 0 });
-        expect(bounds.minY).toBe(-25);
-        expect(bounds.maxX).toBe(100);
-        expect(bounds.minX).toBe(0);
-    });
-});
-
 describe('computeGatheredPositions with tab paths', () => {
     it('should account for tab geometry in layout spacing', () => {
         // Create pieces with tabs that extend 30px beyond the edge
@@ -236,64 +148,5 @@ describe('applyGatheredPositions', () => {
         applyGatheredPositions(groups, new Map([[1, { x: 50, y: 60 }]]));
         expect(groups[0].position).toEqual({ x: 50, y: 60 });
         expect(groups[1].position).toEqual({ x: 300, y: 400 });
-    });
-});
-
-describe('getGroupLocalBounds', () => {
-    it('ignores rotation and returns un-rotated bounds', () => {
-        const piece = makeRectPiece({ id: 1, width: 100, height: 40 });
-        const group = makeGroup(1, 0, 0);
-        group.rotation = 1; // CW 90°
-
-        expect(getGroupLocalBounds(group, [piece])).toEqual({
-            minX: 0,
-            minY: 0,
-            width: 100,
-            height: 40,
-        });
-    });
-});
-
-describe('getGroupVisualBounds', () => {
-    it('matches local bounds at rotation 0', () => {
-        const piece = makeRectPiece({ id: 1, width: 100, height: 40 });
-        const group = makeGroup(1, 0, 0);
-
-        expect(getGroupVisualBounds(group, [piece])).toEqual(
-            getGroupLocalBounds(group, [piece]),
-        );
-    });
-
-    it('swaps width and height at rotation 1 (90° CW)', () => {
-        const piece = makeRectPiece({ id: 1, width: 100, height: 40 });
-        const group = makeGroup(1, 0, 0);
-        group.rotation = 1;
-
-        const bounds = getGroupVisualBounds(group, [piece]);
-
-        expect(bounds.width).toBeCloseTo(40);
-        expect(bounds.height).toBeCloseTo(100);
-    });
-
-    it('swaps width and height at rotation 3 (270° CW)', () => {
-        const piece = makeRectPiece({ id: 1, width: 100, height: 40 });
-        const group = makeGroup(1, 0, 0);
-        group.rotation = 3;
-
-        const bounds = getGroupVisualBounds(group, [piece]);
-
-        expect(bounds.width).toBeCloseTo(40);
-        expect(bounds.height).toBeCloseTo(100);
-    });
-
-    it('keeps width and height at rotation 2 (180°)', () => {
-        const piece = makeRectPiece({ id: 1, width: 100, height: 40 });
-        const group = makeGroup(1, 0, 0);
-        group.rotation = 2;
-
-        const bounds = getGroupVisualBounds(group, [piece]);
-
-        expect(bounds.width).toBeCloseTo(100);
-        expect(bounds.height).toBeCloseTo(40);
     });
 });
