+++ src/game/gather.ts
@@ -9,7 +9,7 @@
  */
 
 import type { Point, Piece, PieceGroup } from '../model/types.js';
-import { rotatePoint } from '../model/helpers.js';
+import { getGroupVisualBounds } from './group-bounds.js';
 
 /** Padding between groups when distributing in the gather layout. */
 export const GATHER_PADDING = 50;
@@ -26,346 +26,6 @@ export interface WorldRect {
     width: number;
     /** Height in world units. */
     height: number;
-}
-
-/**
- * Compute bounding box of piece offsets within a group (group-local space).
- *
- * Returns the min/max of all piece offset coordinates. For a single-piece
- * group at offset (0,0), this returns {minX:0, minY:0, maxX:0, maxY:0}.
- *
- * Note: This uses piece offsets only (not edge geometry). For accurate
- * world-space bounding boxes that include tab shapes, use
- * `getGroupBounds` from pile-detection.
- */
-export function getGroupOffsetBounds(group: PieceGroup): {
-    minX: number;
-    minY: number;
-    maxX: number;
-    maxY: number;
-} {
-    let minX = Infinity;
-    let minY = Infinity;
-    let maxX = -Infinity;
-    let maxY = -Infinity;
-
-    for (const offset of group.pieces.values()) {
-        if (offset.x < minX) {
-            minX = offset.x;
-        }
-
-        if (offset.y < minY) {
-            minY = offset.y;
-        }
-
-        if (offset.x > maxX) {
-            maxX = offset.x;
-        }
-
-        if (offset.y > maxY) {
-            maxY = offset.y;
-        }
-    }
-
-    return { minX, minY, maxX, maxY };
-}
-
-/**
- * Parse an SVG path string and compute the bounding box of all points
- * (including bezier control points) in the path.
- *
- * This produces a conservative bounding box — control points may extend
- * beyond the actual curve, but the result is guaranteed to contain the
- * full path geometry. Good enough for layout spacing.
- */
-export function getPathBounds(
-    path: string,
-    start: Point,
-): { minX: number; minY: number; maxX: number; maxY: number } {
-    let minX = start.x;
-    let minY = start.y;
-    let maxX = start.x;
-    let maxY = start.y;
-
-    let curX = start.x;
-    let curY = start.y;
-
-    function expand(x: number, y: number) {
-        if (x < minX) minX = x;
-        if (y < minY) minY = y;
-        if (x > maxX) maxX = x;
-        if (y > maxY) maxY = y;
-    }
-
-    // Match SVG path commands: a letter followed by numbers/commas/spaces
-    const commandRegex = /([MLCSQTZHVAmlcsqtzhva])\s*([-\d.,eE\s]*)/g;
-    let match: RegExpExecArray | null;
-
-    while ((match = commandRegex.exec(path)) !== null) {
-        const cmd = match[1];
-        const argsStr = match[2].trim();
-        const nums: number[] = [];
-
-        if (argsStr.length > 0) {
-            // Parse numbers, handling negative signs and decimals
-            const numRegex = /[-+]?(?:\d+\.?\d*|\.\d+)(?:[eE][-+]?\d+)?/g;
-            let numMatch: RegExpExecArray | null;
-            while ((numMatch = numRegex.exec(argsStr)) !== null) {
-                nums.push(parseFloat(numMatch[0]));
-            }
-        }
-
-        switch (cmd) {
-            case 'M':
-                for (let i = 0; i + 1 < nums.length; i += 2) {
-                    curX = nums[i];
-                    curY = nums[i + 1];
-                    expand(curX, curY);
-                }
-                break;
-            case 'm':
-                for (let i = 0; i + 1 < nums.length; i += 2) {
-                    curX += nums[i];
-                    curY += nums[i + 1];
-                    expand(curX, curY);
-                }
-                break;
-            case 'L':
-                for (let i = 0; i + 1 < nums.length; i += 2) {
-                    curX = nums[i];
-                    curY = nums[i + 1];
-                    expand(curX, curY);
-                }
-                break;
-            case 'l':
-                for (let i = 0; i + 1 < nums.length; i += 2) {
-                    curX += nums[i];
-                    curY += nums[i + 1];
-                    expand(curX, curY);
-                }
-                break;
-            case 'H':
-                for (const n of nums) {
-                    curX = n;
-                    expand(curX, curY);
-                }
-                break;
-            case 'h':
-                for (const n of nums) {
-                    curX += n;
-                    expand(curX, curY);
-                }
-                break;
-            case 'V':
-                for (const n of nums) {
-                    curY = n;
-                    expand(curX, curY);
-                }
-                break;
-            case 'v':
-                for (const n of nums) {
-                    curY += n;
-                    expand(curX, curY);
-                }
-                break;
-            case 'C':
-                for (let i = 0; i + 5 < nums.length; i += 6) {
-                    expand(nums[i], nums[i + 1]);
-                    expand(nums[i + 2], nums[i + 3]);
-                    curX = nums[i + 4];
-                    curY = nums[i + 5];
-                    expand(curX, curY);
-                }
-                break;
-            case 'c':
-                for (let i = 0; i + 5 < nums.length; i += 6) {
-                    expand(curX + nums[i], curY + nums[i + 1]);
-                    expand(curX + nums[i + 2], curY + nums[i + 3]);
-                    curX += nums[i + 4];
-                    curY += nums[i + 5];
-                    expand(curX, curY);
-                }
-                break;
-            case 'S':
-                for (let i = 0; i + 3 < nums.length; i += 4) {
-                    expand(nums[i], nums[i + 1]);
-                    curX = nums[i + 2];
-                    curY = nums[i + 3];
-                    expand(curX, curY);
-                }
-                break;
-            case 's':
-                for (let i = 0; i + 3 < nums.length; i += 4) {
-                    expand(curX + nums[i], curY + nums[i + 1]);
-                    curX += nums[i + 2];
-                    curY += nums[i + 3];
-                    expand(curX, curY);
-                }
-                break;
-            case 'Q':
-                for (let i = 0; i + 3 < nums.length; i += 4) {
-                    expand(nums[i], nums[i + 1]);
-                    curX = nums[i + 2];
-                    curY = nums[i + 3];
-                    expand(curX, curY);
-                }
-                break;
-            case 'q':
-                for (let i = 0; i + 3 < nums.length; i += 4) {
-                    expand(curX + nums[i], curY + nums[i + 1]);
-                    curX += nums[i + 2];
-                    curY += nums[i + 3];
-                    expand(curX, curY);
-                }
-                break;
-            case 'T':
-                for (let i = 0; i + 1 < nums.length; i += 2) {
-                    curX = nums[i];
-                    curY = nums[i + 1];
-                    expand(curX, curY);
-                }
-                break;
-            case 't':
-                for (let i = 0; i + 1 < nums.length; i += 2) {
-                    curX += nums[i];
-                    curY += nums[i + 1];
-                    expand(curX, curY);
-                }
-                break;
-            case 'A':
-                for (let i = 0; i + 6 < nums.length; i += 7) {
-                    // Include the arc endpoint; radii are harder to bound
-                    // but including the endpoint is a reasonable approximation
-                    curX = nums[i + 5];
-                    curY = nums[i + 6];
-                    expand(curX, curY);
-                }
-                break;
-            case 'a':
-                for (let i = 0; i + 6 < nums.length; i += 7) {
-                    curX += nums[i + 5];
-                    curY += nums[i + 6];
-                    expand(curX, curY);
-                }
-                break;
-            case 'Z':
-            case 'z':
-                // Close path — return to start
-                curX = start.x;
-                curY = start.y;
-                break;
-        }
-    }
-
-    return { minX, minY, maxX, maxY };
-}
-
-/**
- * Compute the bounding box of a group in its un-rotated local space by
- * examining the actual SVG shape geometry of its pieces. Piece offsets
- * live in un-rotated local coordinates, and so do the bounds this returns.
- *
- * Includes bezier control points from edge paths to account for tab
- * geometry that extends beyond the start/end corner vertices.
- *
- * Use this (not `getGroupVisualBounds`) when doing rotation pivot math or
- * anywhere else you need rotation-invariant bounds.
- */
-export function getGroupLocalBounds(
-    group: PieceGroup,
-    pieces: ReadonlyArray<Readonly<Piece>>,
-): { minX: number; minY: number; width: number; height: number } {
-    let minX = Infinity;
-    let minY = Infinity;
-    let maxX = -Infinity;
-    let maxY = -Infinity;
-
-    for (const [pieceId, offset] of group.pieces) {
-        const piece = pieces.find(p => p.id === pieceId);
-        if (!piece) continue;
-
-        for (const edge of piece.edges) {
-            // Always include start/end points
-            const points = [edge.start, edge.end];
-            for (const pt of points) {
-                const wx = offset.x + pt.x;
-                const wy = offset.y + pt.y;
-                if (wx < minX) minX = wx;
-                if (wy < minY) minY = wy;
-                if (wx > maxX) maxX = wx;
-                if (wy > maxY) maxY = wy;
-            }
-
-            // Include path geometry (control points, curve endpoints)
-            if (edge.path) {
-                const pathBounds = getPathBounds(edge.path, edge.start);
-                const pts = [
-                    { x: pathBounds.minX, y: pathBounds.minY },
-                    { x: pathBounds.maxX, y: pathBounds.maxY },
-                ];
-                for (const pt of pts) {
-                    const wx = offset.x + pt.x;
-                    const wy = offset.y + pt.y;
-                    if (wx < minX) minX = wx;
-                    if (wy < minY) minY = wy;
-                    if (wx > maxX) maxX = wx;
-                    if (wy > maxY) maxY = wy;
-                }
-            }
-        }
-    }
-
-    if (!isFinite(minX)) {
-        return { minX: 0, minY: 0, width: 0, height: 0 };
-    }
-
-    return { minX, minY, width: maxX - minX, height: maxY - minY };
-}
-
-/**
- * Compute the bounding box of a group as it actually renders, accounting
- * for `group.rotation`. Returned coordinates are offsets from the group's
- * `position` (pre-translation) in rotated local space, so
- * `group.position.x + bounds.minX` is the world-space left edge.
- *
- * Use this for layout, gather packing, or anywhere the rendered footprint
- * matters. For rotation pivot math use `getGroupLocalBounds` instead.
- */
-export function getGroupVisualBounds(
-    group: PieceGroup,
-    pieces: ReadonlyArray<Readonly<Piece>>,
-): { minX: number; minY: number; width: number; height: number } {
-    const local = getGroupLocalBounds(group, pieces);
-
-    if (group.rotation === 0) {
-        return local;
-    }
-
-    if (local.width === 0 && local.height === 0) {
-        return local;
-    }
-
-    // Rotate the four local-space corners and recompute the AABB.
-    const corners: Point[] = [
-        { x: local.minX, y: local.minY },
-        { x: local.minX + local.width, y: local.minY },
-        { x: local.minX + local.width, y: local.minY + local.height },
-        { x: local.minX, y: local.minY + local.height },
-    ];
-
-    let minX = Infinity;
-    let minY = Infinity;
-    let maxX = -Infinity;
-    let maxY = -Infinity;
-    for (const c of corners) {
-        const r = rotatePoint(c, group.rotation);
-        if (r.x < minX) minX = r.x;
-        if (r.y < minY) minY = r.y;
-        if (r.x > maxX) maxX = r.x;
-        if (r.y > maxY) maxY = r.y;
-    }
-
-    return { minX, minY, width: maxX - minX, height: maxY - minY };
 }
 
 /**
